test(e2e): Try to fix issues by refactor
Try to see if putting tests in different packages affects results.
